Move existing 3PID settings UX to account directory
