+++ src/components/views/settings/account/PhoneNumbers.js
@@ -17,14 +17,14 @@ limitations under the License.
 
 import React from 'react';
 import PropTypes from 'prop-types';
-import {_t} from "../../../languageHandler";
-import MatrixClientPeg from "../../../MatrixClientPeg";
-import Field from "../elements/Field";
-import AccessibleButton from "../elements/AccessibleButton";
-import AddThreepid from "../../../AddThreepid";
-import CountryDropdown from "../auth/CountryDropdown";
-const sdk = require('../../../index');
-const Modal = require("../../../Modal");
+import {_t} from "../../../../languageHandler";
+import MatrixClientPeg from "../../../../MatrixClientPeg";
+import Field from "../../elements/Field";
+import AccessibleButton from "../../elements/AccessibleButton";
+import AddThreepid from "../../../../AddThreepid";
+import CountryDropdown from "../../auth/CountryDropdown";
+const sdk = require('../../../../index');
+const Modal = require("../../../../Modal");
 
 /*
 TODO: Improve the UX for everything in here.
